+++ app/exerlog/lib/UI/login_screen/auth_data.dart
@@ -1,0 +1,7 @@
+class AuthData {
+  String password;
+  String email;
+  String? confirmPassword;
+
+  AuthData({required this.email, required this.password, this.confirmPassword});
+}
