--- app/exerlog/lib/UI/login_screen/login_data.dart
@@ -1,8 +1,0 @@
-class LoginData {
-  String password;
-  String email;
-  String? confirmPassword;
-
-  LoginData(
-      {required this.email, required this.password, this.confirmPassword});
-}
