+++ app/exerlog/lib/UI/login_screen/signup_form.dart
@@ -1,9 +1,9 @@
 import 'package:exerlog/UI/global.dart';
-import 'package:exerlog/UI/login_screen/login_data.dart';
+import 'package:exerlog/UI/login_screen/auth_data.dart';
 import 'package:flutter/material.dart';
 
 class SignupForm extends StatefulWidget {
-  final LoginData loginData;
+  final AuthData loginData;
 
   const SignupForm(this.loginData);
